fix: adding textInput videos
In the process of adding text input videos, I realized the "magic" of how the videos were being output on the page was not going to allow us to do native videos with different states like Debbie was hoping. Ithus reworked the logic of how these output based on doing a "fuzzier" search on the slug and the category (native or web). This definitely needs a review and clean up and another round.